#309 Move DNS to infrastructure
