+++ server/internal/cluster/domain/port/dns.go
@@ -1,0 +1,8 @@
+package port
+
+import "context"
+
+type DNS interface {
+	WaitForHost(ctx context.Context, host string, expectedIP string) error
+	AllHostsResolve(hosts []string, expectedIP string) bool
+}
