+++ server/internal/cluster/infrastructure/dns/module.go
@@ -1,0 +1,10 @@
+package dns
+
+import "go.uber.org/fx"
+
+var Module = fx.Module(
+	"dns",
+	fx.Provide(
+		NewResolver,
+	),
+)
